requirements.txt
fastapi
uvicorn
requests
pyjwt
cryptography